Fix hash collision problem for Shape object
The problem is that arguments of Shape object are assumed to be
sorted according to their hash values. But if some arguments have
the same hash, than their order become undefined. But equality
procedure relies on the order. This approach is modified to take
into account possible hash collisions. If some arguments have
equal hashes, than all possible permutations are handled.