build: reorganizing access and building of nlopt
when built locally everything works fine because libnlopt is installed to site lib directory to be linked with
geometry module. At the same time it is copied to build package and is avaiable in the wheel. Besides there's the case with 'development' install in that case we have to search it in site's lib